Replace Booster model with Badge in the domain
Sorry, this commit is a bit of a mess; you can see the reason why by looking at the commit dates. It's definitely not an atomic commit, and there are unrelated changes.

This commit introduces the concept of badges, which more or less used to be boosters. The reasoning was that if a game has boosters, you can craft them, but the introduction of badges represents the domain better.