Create class-based model factory for games
Class-based model factories are based on this blog post:

https://tighten.co/blog/tidy-up-your-tests-with-class-based-model-factories

I encountered a problem when trying to use it as a real-time facade: the instance is reused, so its properties are too. So, for example, if I created some games with an owner and after I tried to create one without one, it actually reused and assigned the previous user as owner.

https://github.com/laravel/ideas/issues/1088