Revert crafting cost property from games to badges
Four commits ago, I changed my mind about where the crafting cost goes.
My mental model about boosters was that they were the individual
boosters instead of the generic properties for a booster, so it didn't
make sense to have the crafting cost in them, because you couldn't craft
one without it.

But this was wrong, of course. If I need to "create" many boosters for
some reason in the future (for example, the user's inventory) I can just use a
pivot table.